+++ test/Spec.hs
@@ -5,6 +5,7 @@ import Test.Hspec
 
 import qualified File.MapIOSpec
 import qualified Game.GridUtilsSpec
+import qualified Game.AutoplaySpec
 import qualified Game.LogicSpec
 import qualified Game.StateSpec
 import qualified UI.DrawSpec
@@ -17,6 +18,7 @@ main = hspec $ do
   describe "Game.State"     Game.StateSpec.spec
   describe "Game.Logic"     Game.LogicSpec.spec
   describe "File.MapIO"     File.MapIOSpec.spec
+  describe "Game.Autoplay"  Game.AutoplaySpec.spec
   describe "UI.Draw"        UI.DrawSpec.spec
   describe "UI.Layout"      UI.LayoutSpec.spec
   describe "UI.Screen"      UI.ScreenSpec.spec